Split DI for REST API and networking
