+++ data/api/src/main/java/com/alvindizon/panahon/api/di/ApiModule.kt
@@ -1,0 +1,36 @@
+package com.alvindizon.panahon.api.di
+
+import com.alvindizon.panahon.api.OpenWeatherApi
+import com.squareup.moshi.Moshi
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import okhttp3.OkHttpClient
+import retrofit2.Retrofit
+import retrofit2.converter.moshi.MoshiConverterFactory
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object ApiModule {
+
+    private const val OPENWEATHER_URL = "https://api.openweathermap.org/"
+
+    @Provides
+    @Singleton
+    fun provideMoshi(): Moshi = Moshi.Builder().build()
+
+    @Provides
+    @Singleton
+    fun provideRetrofit(okHttpClient: OkHttpClient, moshi: Moshi): Retrofit = Retrofit.Builder()
+        .baseUrl(OPENWEATHER_URL)
+        .addConverterFactory(MoshiConverterFactory.create(moshi))
+        .client(okHttpClient)
+        .build()
+
+    @Provides
+    @Singleton
+    fun provideOpenWeatherApi(retrofit: Retrofit): OpenWeatherApi =
+        retrofit.create(OpenWeatherApi::class.java)
+}
